Moved unit tests to appropriate folder
